Perfundimi i Backend

+++ LibrariaProjekt.Server/Controllers/UserController.cs
@@ -4,65 +4,66 @@ using LibrariaProjekt.Server.Repositories;
 
 namespace LibrariaProjekt.Server.Controllers
 {
-    [Route("api/[controller]")]
-    [ApiController]
     public class UserController : Controller
     {
         private readonly IUserRepository _userRepository;
 
-        public UserController(IUserRepository
-            userRepository)
+        public UserController(IUserRepository userRepository)
         {
             _userRepository = userRepository;
         }
 
-
         public IActionResult Index()
         {
-            var entries = _userRepository.GetAll().ToList();
+            var entries = _userRepository.GetAll();
             return View(entries);
         }
 
         [HttpGet]
         public IActionResult Create()
         {
-
-
             return View();
         }
 
         [HttpPost]
-        public IActionResult Create(User obj)
+      
+        public IActionResult Create(User user)
         {
-            _userRepository.Insert(obj);
+            _userRepository.Insert(user);
             return RedirectToAction("Index");
         }
-
 
         [HttpGet]
         public IActionResult Edit(int id)
         {
-            User s1 = _userRepository.GetById(id);
-            return View(s1);
+            User user = _userRepository.GetById(id);
+            return View(user);
         }
 
         [HttpPost]
+        
         public IActionResult Edit(User user)
         {
             _userRepository.Update(user);
             return RedirectToAction("Index");
         }
+
         [HttpGet]
         public IActionResult Delete(int id)
         {
-            User s1 = _userRepository.GetById(id);
-            return View(s1);
+            User user = _userRepository.GetById(id);
+            return View(user);
         }
 
         [HttpPost]
-        public IActionResult Delete(User user)
+      
+        public IActionResult DeleteConfirmed(int id)
         {
-            _userRepository.Delete(user);
+            User user = _userRepository.GetById(id);
+            if (user != null)
+            {
+                _userRepository.Delete(user);
+            }
             return RedirectToAction("Index");
         }
     }
